+++ src/frontend/web_application/src/store/configure-store.js
@@ -1,5 +1,4 @@
 import { createStore, applyMiddleware } from 'redux';
-import { createScrollMiddleware } from 'react-redux-scroll';
 import rootReducer from './reducer';
 import applicationMiddleware from './middlewares/application-middleware';
 import axiosMiddleware from './middlewares/axios-middleware';
@@ -34,7 +33,6 @@ const middlewares = [
   tabsMiddleware,
   tagsMiddleware,
   thunkMiddleware,
-  createScrollMiddleware(),
 ];
 
 if (CALIOPEN_ENV === 'development' || CALIOPEN_ENV === 'staging') {
@@ -43,6 +41,7 @@ if (CALIOPEN_ENV === 'development' || CALIOPEN_ENV === 'staging') {
 }
 
 if (BUILD_TARGET === 'browser') {
+  middlewares.push(require('react-redux-scroll').createScrollMiddleware()); // eslint-disable-line
   middlewares.push(require('./middlewares/openpgp-middleware').default); // eslint-disable-line
 }